Split shop tests, add given, when, then
